move some more client stuff around.
