+++ packages/finch-graphql/src/client/client.test.ts
@@ -1,6 +1,6 @@
 import { queryApi } from './client';
 import gql from 'graphql-tag';
-import { FinchMessageKey } from './types';
+import { FinchMessageKey } from '../types';
 
 describe('queryApi', () => {
   it('should send and message to the background script', async () => {
