+++ packages/finch-graphql/src/client/hooks/useMutation.test.ts
@@ -1,8 +1,8 @@
 import React from 'react';
 import { renderHook, act } from '@testing-library/react-hooks';
 import gql from 'graphql-tag';
-import { FinchMessageKey } from '../types';
-import { ExtensionProvider } from './ExtensionProvider';
+import { FinchMessageKey } from '../../types';
+import { FinchProvider } from './FinchProvider';
 import { useMutation } from './useMutation';
 
 const testDoc = gql`
@@ -66,7 +66,7 @@ describe('useMutation', () => {
     const wrapper = renderHook(() => useMutation(testDoc), {
       wrapper: ({ children }) => {
         // @ts-ignore
-        return React.createElement(ExtensionProvider, { children, id: 'foo' });
+        return React.createElement(FinchProvider, { children, id: 'foo' });
       },
     });
