+++ packages/finch-graphql/src/client/index.ts
@@ -1,1 +1,4 @@
 export { queryApi } from './client';
+export { FinchProvider, useFinchClient } from './hooks/FinchProvider';
+export { useQuery } from './hooks/useQuery';
+export { useMutation } from './hooks/useMutation';
