+++ packages/finch-graphql/src/index.ts
@@ -2,7 +2,6 @@ export * from './background';
 export * from './client';
 export * from './types';
 export * from './external';
-export { ExtensionProvider, useExtension } from './hooks/ExtensionProvider';
-export { useQuery } from './hooks/useQuery';
-export { useMutation } from './hooks/useMutation';
+export { useQuery } from './client/hooks/useQuery';
+export { useMutation } from './client/hooks/useMutation';
 export { FinchError } from './errors/FinchError';
